backend/requirements.txt
anyio>=4.12.1
blinker>=1.9.0
certifi>=2026.1.4
charset-normalizer>=3.4.4
click>=8.3.1
colorama>=0.4.6
dnspython>=2.8.0
faiss-cpu>=1.13.2
filelock>=3.20.3
Flask>=3.1.2
flask-cors>=6.0.2
Flask-JWT-Extended>=4.7.1
fsspec>=2026.2.0
h11>=0.16.0
hf-xet>=1.2.0
httpcore>=1.0.9
httpx>=0.28.1
huggingface_hub>=1.4.1
idna>=3.11
itsdangerous>=2.2.0
Jinja2>=3.1.6
joblib>=1.5.3
MarkupSafe>=3.0.3
mpmath>=1.3.0
networkx>=3.6.1
numpy>=2.4.2
packaging>=26.0
pandas>=3.0.0
PyJWT>=2.11.0
Flask-PyMongo>=2.3.0
python-dateutil>=2.9.0.post0
python-dotenv>=1.2.1
PyYAML>=6.0.3
regex>=2026.1.15
requests>=2.32.5
safetensors>=0.7.0
scikit-learn>=1.8.0
scipy>=1.17.0
sentence-transformers>=5.2.2
setuptools>=82.0.0
shellingham>=1.5.4
six>=1.17.0
soupsieve>=2.8.3
sympy>=1.14.0
threadpoolctl>=3.6.0
tokenizers>=0.22.2
torch>=2.10.0
tqdm>=4.67.3
transformers>=5.1.0
typer-slim>=0.21.1
typing_extensions>=4.15.0
tzdata>=2025.3
urllib3>=2.6.3
Werkzeug>=3.1.5
gunicorn>=21.2.0
deep-translator>=1.11.4
